cocoaui, pltbrowser: fix int overflow bug resulting in negative values in total playback time calculation (fixes #3270)

+++ plugins/cocoaui/DesignMode/Widgets/PlaylistBrowser/PlaylistBrowserViewController.m
@@ -270,10 +270,10 @@ extern DB_functions_t *deadbeef;
     }
     else if ([tableColumn.identifier isEqualToString:@"Duration"]) {
         float pl_totaltime = deadbeef->plt_get_totaltime (plt);
-        int daystotal = (int)pl_totaltime / (3600*24);
-        int hourtotal = ((int)pl_totaltime / 3600) % 24;
-        int mintotal = ((int)pl_totaltime/60) % 60;
-        int sectotal = ((int)pl_totaltime) % 60;
+        int daystotal = (int)(int64_t)pl_totaltime / (3600 * 24);
+        int hourtotal = (int)((int64_t)pl_totaltime / 3600) % 24;
+        int mintotal = (int)((int64_t)pl_totaltime / 60) % 60;
+        int sectotal = (int)((int64_t)pl_totaltime) % 60;
 
         char totaltime_str[512] = "";
         if (daystotal == 0) {
